VMM/IEM: Fixed incorrect type for the fAlignMaskAndClt parameter of iemNativeEmitTlbLookup(), introduced in r163594. bugref:10687
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@105167 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/include/IEMN8veRecompilerTlbLookup.h
@@ -328,7 +328,7 @@ DECLASM(void) iemNativeHlpAsmSafeWrapCheckTlbLookup(void);
 template<bool const a_fDataTlb, bool const a_fNoReturn = false>
 DECL_INLINE_THROW(uint32_t)
 iemNativeEmitTlbLookup(PIEMRECOMPILERSTATE pReNative, uint32_t off, IEMNATIVEEMITTLBSTATE const * const pTlbState,
-                       uint8_t iSegReg, uint8_t cbMem, uint8_t fAlignMaskAndCtl, uint32_t fAccess,
+                       uint8_t iSegReg, uint8_t cbMem, uint32_t fAlignMaskAndCtl, uint32_t fAccess,
                        uint32_t idxLabelTlbLookup, uint32_t idxLabelTlbMiss, uint8_t idxRegMemResult,
                        uint8_t offDisp = 0)
 {
